requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
langchain-google-genai
langchain-community
langchain-text-splitters
langchain-chroma
chromadb
docx2txt
